feat(): copy job model into app
add loggeable concern too
