+++ lib/generators/active_job_log/install/templates/job_model.rb
@@ -1,0 +1,5 @@
+module ActiveJobLog
+  class Job < ApplicationRecord
+    include Loggeable
+  end
+end
