+++ spec/dummy/app/models/active_job_log/job.rb
@@ -1,0 +1,5 @@
+module ActiveJobLog
+  class Job < ApplicationRecord
+    include Loggeable
+  end
+end
